# apps/backend/src/modules/memberships/dto/organisation-membership.dto.ts
import { OrganisationMembershipSchema } from '@repo/contracts/memberships';
import { createZodDto } from 'nestjs-zod';
export class OrganisationMembershipDto extends createZodDto(OrganisationMembershipSchema) {}