# packages/contracts/openapi/openapi.yaml
openapi: 3.0.0
paths:
...
  /api/memberships:
    get:
      operationId: MembershipsController_list
      parameters:
        - name: userId
          required: false
          in: query
          schema:
            minLength: 1
            type: string
        - name: organisationId
          required: false
          in: query
          schema:
            format: uuid
            pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListMembershipsResponseDto"
        "401":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "403":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
      security: &a7
        - session: []
      summary: List memberships. Non-sysadmins may only filter by their own userId.
      tags: &a8
        - memberships
    post:
      operationId: MembershipsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMembershipDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganisationMembershipDto"
        "400":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "401":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "403":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "409":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
      security: *a7
      summary: Create a membership (sysadmin only).
      tags: *a8
  /api/memberships/{id}:
    patch:
      operationId: MembershipsController_update
      parameters:
        - name: id
          required: true
          in: path
          description: Membership UUID.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMembershipDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganisationMembershipDto"
        "400":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "401":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "403":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "404":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
      security: *a7
      summary: Change a membership’s role (sysadmin only).
      tags: *a8
    delete:
      operationId: MembershipsController_delete
      parameters:
        - name: id
          required: true
          in: path
          description: Membership UUID.
          schema:
            type: string
      responses:
        "204":
          description: Membership deleted.
        "401":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "403":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "404":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
      security: *a7
      summary: Delete a membership (sysadmin only).
      tags: *a8
...
info:
  title: taidohub API
  description: HTTP boundary for taidohub — schemas derived from
    `@repo/contracts`, enforced at runtime by `ZodValidationPipe` and documented
    here.
  version: 0.0.0
  contact: {}
...
servers:
  - url: http://localhost:3001
    description: Local
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: better-auth.session_token
    bearer:
      scheme: bearer
      bearerFormat: session-token
      type: http
  schemas:
...
    ErrorEnvelopeDto:
      type: object
      properties:
        error:
          $ref: "#/$defs/ErrorPayload"
          type: ""
          x-nestjs_zod-parent-has-refs: true
          x-nestjs_zod-empty-type: true
          x-nestjs_zod-$defs:
            ErrorPayload:
              type: object
              properties:
                code:
                  type: string
                  description: Stable, machine-readable error code (e.g. `VALIDATION_FAILED`).
                message:
                  type: string
                  description: Human-readable error message safe to surface to end users.
                details:
                  description: Optional structured context — e.g. per-field validation errors.
              required:
                - code
                - message
              description: The body of a uniform error envelope.
              example:
                code: NOT_FOUND
                message: The requested resource could not be found.
          x-nestjs_zod-parent-metadata:
            description: The uniform error envelope returned for every non-2xx response from
              the API.
            example:
              error:
                code: NOT_FOUND
                message: The requested resource could not be found.
      required:
        - error
...
    ListMembershipsResponseDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/$defs/OrganisationMembership"
          x-nestjs_zod-parent-has-refs: true
          x-nestjs_zod-$defs: &a20
            OrganisationMembership:
              type: object
              properties:
                id:
                  type: string
                  format: uuid
                  pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                userId:
                  type: string
                  minLength: 1
                organisationId:
                  type: string
                  format: uuid
                  pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                role:
                  $ref: "#/$defs/MembershipRole"
                createdAt:
                  type: string
                  format: date-time
                  pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                updatedAt:
                  type: string
                  format: date-time
                  pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
              required:
                - id
                - userId
                - organisationId
                - role
                - createdAt
                - updatedAt
              description: A `(user, organisation, role)` binding.
              example:
                id: 7d3a2e0e-2e8c-4b7a-9a6e-1f9d1e54b8f5
                userId: u-admin
                organisationId: 0e2e8c4b-7a9a-46e1-9d1e-54b8f7d3a2e0
                role: orgadmin
                createdAt: 2026-05-18T08:00:00.000Z
                updatedAt: 2026-05-18T08:00:00.000Z
            MembershipRole:
              type: string
              enum:
                - orgadmin
                - instructor
              description: Org-scoped role on an `organisation_membership` row.
              example: orgadmin
        total:
          type: integer
          minimum: 0
          maximum: 9007199254740991
          x-nestjs_zod-parent-has-refs: true
          x-nestjs_zod-$defs: *a20
      required:
        - data
        - total
    CreateMembershipDto:
      type: object
      properties:
        userId:
          type: string
          minLength: 1
          x-nestjs_zod-parent-has-refs: true
          x-nestjs_zod-$defs: &a21
            MembershipRole:
              type: string
              enum:
                - orgadmin
                - instructor
              description: Org-scoped role on an `organisation_membership` row.
              example: orgadmin
        organisationId:
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          x-nestjs_zod-parent-has-refs: true
          x-nestjs_zod-$defs: *a21
        role:
          $ref: "#/$defs/MembershipRole"
          type: ""
          x-nestjs_zod-parent-has-refs: true
          x-nestjs_zod-empty-type: true
          x-nestjs_zod-$defs: *a21
      required:
        - userId
        - organisationId
        - role
    OrganisationMembershipDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          x-nestjs_zod-parent-has-refs: true
          x-nestjs_zod-$defs: &a22
            MembershipRole:
              type: string
              enum:
                - orgadmin
                - instructor
              description: Org-scoped role on an `organisation_membership` row.
              example: orgadmin
          x-nestjs_zod-parent-metadata:
            description: A `(user, organisation, role)` binding.
            example: &a23
              id: 7d3a2e0e-2e8c-4b7a-9a6e-1f9d1e54b8f5
              userId: u-admin
              organisationId: 0e2e8c4b-7a9a-46e1-9d1e-54b8f7d3a2e0
              role: orgadmin
              createdAt: 2026-05-18T08:00:00.000Z
              updatedAt: 2026-05-18T08:00:00.000Z
        userId:
          type: string
          minLength: 1
          x-nestjs_zod-parent-has-refs: true
          x-nestjs_zod-$defs: *a22
          x-nestjs_zod-parent-metadata:
            description: A `(user, organisation, role)` binding.
            example: *a23
        organisationId:
          type: string
          format: uuid
          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
          x-nestjs_zod-parent-has-refs: true
          x-nestjs_zod-$defs: *a22
          x-nestjs_zod-parent-metadata:
            description: A `(user, organisation, role)` binding.
            example: *a23
        role:
          $ref: "#/$defs/MembershipRole"
          type: ""
          x-nestjs_zod-parent-has-refs: true
          x-nestjs_zod-empty-type: true
          x-nestjs_zod-$defs: *a22
          x-nestjs_zod-parent-metadata:
            description: A `(user, organisation, role)` binding.
            example: *a23
        createdAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
          x-nestjs_zod-parent-has-refs: true
          x-nestjs_zod-$defs: *a22
          x-nestjs_zod-parent-metadata:
            description: A `(user, organisation, role)` binding.
            example: *a23
        updatedAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
          x-nestjs_zod-parent-has-refs: true
          x-nestjs_zod-$defs: *a22
          x-nestjs_zod-parent-metadata:
            description: A `(user, organisation, role)` binding.
            example: *a23
      required:
        - id
        - userId
        - organisationId
        - role
        - createdAt
        - updatedAt
    UpdateMembershipDto:
      type: object
      properties:
        role:
          $ref: "#/$defs/MembershipRole"
          type: ""
          x-nestjs_zod-parent-has-refs: true
          x-nestjs_zod-empty-type: true
          x-nestjs_zod-$defs:
            MembershipRole:
              type: string
              enum:
                - orgadmin
                - instructor
              description: Org-scoped role on an `organisation_membership` row.
              example: orgadmin
      required:
        - role
...